add new endpoints

+++ backend/src/main/java/com/nocountry/backend/controller/CourseController.java
@@ -16,21 +16,25 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.nocountry.backend.dto.CourseDto;
-import com.nocountry.backend.dto.StudentDto;
 import com.nocountry.backend.service.ICourseService;
 
 import lombok.RequiredArgsConstructor;
 
 @RestController
-@RequestMapping("api/v1")
+@RequestMapping("/api")
 @RequiredArgsConstructor
 public class CourseController {
 
     private final ICourseService service;
 
-    @GetMapping("/courses/{id}")
-    public ResponseEntity<Optional<CourseDto>> getCourseById(@PathVariable Long courseId) {
-        return new ResponseEntity<>(service.getCourseById(courseId), HttpStatus.OK);
+    @GetMapping("/courses/{courseId}")
+    public ResponseEntity<Optional<CourseDto>> getCourse(@PathVariable Long courseId) {
+        Optional<CourseDto> courseDto = service.getCourse(courseId);
+        if (courseDto.isPresent()) {
+            return new ResponseEntity<>(courseDto, HttpStatus.OK);
+        } else {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
     }
 
     @GetMapping("/admin/courses/all")
@@ -45,30 +49,34 @@ public class CourseController {
         return new ResponseEntity<>(service.createCourse(courseDto), HttpStatus.CREATED);
     }
 
-    @PutMapping("/admin/courses/update/{id}")
+    @PutMapping("/admin/courses/update/{courseId}")
     @PreAuthorize("hasRole('ADMIN')")
     public ResponseEntity<CourseDto> updateCourse(@RequestBody CourseDto courseDto, @PathVariable Long courseId) {
         return new ResponseEntity<>(service.updateCourse(courseDto, courseId), HttpStatus.ACCEPTED);
     }
 
-    @DeleteMapping("/admin/courses/delete/{id}")
+    @DeleteMapping("/admin/courses/delete/{courseId}")
     @PreAuthorize("hasRole('ADMIN')")
     public ResponseEntity<String> deleteCourse(@PathVariable Long courseId) {
-        service.deleteCourse(courseId);
-        return new ResponseEntity<>("Course successfully deleted", HttpStatus.ACCEPTED);
+        if (service.getCourse(courseId).isPresent()) {
+            service.deleteCourse(courseId);
+            return new ResponseEntity<>("Course successfully deleted", HttpStatus.ACCEPTED);
+        } else {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
     }
 
-    @GetMapping("/admin/courses/{id}/students/all")
+    @PutMapping("/admin/courses/{courseId}/students/add/{studentId}")
     @PreAuthorize("hasRole('ADMIN')")
-    public ResponseEntity<List<StudentDto>> getStudentsByCourseId(@PathVariable Long courseId) {
-        return new ResponseEntity<>(service.getStudentsByCourseId(courseId), HttpStatus.OK);
-
+    public ResponseEntity<String> addStudentToCourse(@PathVariable Long courseId, @PathVariable Long studentId) {
+        service.addStudentToCourse(courseId, studentId);
+        return new ResponseEntity<>("Student successfully added to course", HttpStatus.OK);
     }
 
-    @PostMapping("/admin/courses/{id}/students/add")
+    @PutMapping("/admin/courses/{courseId}/teachers/add/{teacherId}")
     @PreAuthorize("hasRole('ADMIN')")
-    public ResponseEntity<String> addStudentToCourse(@PathVariable Long courseId, @RequestBody StudentDto student) {
-        service.addStudentToCourse(courseId, student);
-        return new ResponseEntity<>("Student successfully added to course", HttpStatus.CREATED);
+    public ResponseEntity<String> addTeacherToCourse(@PathVariable Long courseId, @PathVariable Long teacherId) {
+        service.addTeacherToCourse(courseId, teacherId);
+        return new ResponseEntity<>("Teacher successfully added to course", HttpStatus.OK);
     }
 }
